chore: Rework integration tests (#209)
